Activate beartype claw at lcm.__init__ + fix surfaced drift
Moves the `beartype_package(...)` registrations from `tests/conftest.py`
into `lcm/__init__.py` so the claw actually instruments `lcm.grids`,
`lcm.shocks`, and `lcm.params` at first import. The conftest version
ran AFTER `from lcm._beartype_conf import ...` triggered
`lcm/__init__.py`, which eagerly imports every targeted submodule —
the claw was a no-op on every prior run.

Live-claw drift fixed:

- `Piece` `n_points` cumulative sum: explicit `dtype=jnp.int32` on the
  `_piece_n_points` array + `sum(dtype=jnp.int32)` so the property
  honours its `ScalarInt` annotation under `jax_enable_x64=True`.
- `_ShockGrid.compute_gridpoints` / `compute_transition_probs` /
  `Tauchen._innovation_variance`: kwargs accept `float | FloatND |
  IntND`; runtime params (e.g. `n_std=2`) flow through as 0-d int
  arrays after pylcm's canonical-dtype cast.
- `UniformContinuousGrid` / `LinSpacedGrid` / `LogSpacedGrid` /
  `_ShockGrid` `get_coordinate`: accept Python `float` / `int`
  alongside `ScalarFloat | FloatND`. Same widening on
  `get_irreg_coordinate`.
- `process_params(params_template=...)`: widen to
  `Mapping[RegimeName, Mapping[str, object]]` since real callers and
  test mocks both use plain dicts at the inner levels (the strict
  `MappingProxyType[...]`-only alias rejected them).
- `_ParamsLeaf`: add `np.ndarray` so numpy-array params (used in
  dtype-invariant tests) pass the perimeter check before pylcm casts
  them.
- `create_regime_params_template(regime: ...)` and helpers,
  `create_params_template(internal_regimes: ...)`: typed as `Any` so
  duck-typed test mocks satisfy the signature (the functions
  themselves only touch a small structural interface — `states`,
  `actions`, `functions`, `regime_params_template`).
- `test_as_leaf_rejects_int` now expects `InvalidParamsError` (the
  configured `PARAMS_CONF` exception), matching the claw-routed
  violation; the in-function `TypeError` raise no longer fires
  because beartype rejects the argument first.

Drops the now-redundant claw setup from `tests/conftest.py` and
strips the `# ty: ignore[invalid-argument-type]` / `[no-matching-
overload]` markers the widenings make unnecessary.

+++ src/lcm/grids/piecewise.py
@@ -87,7 +87,7 @@ class PiecewiseLinSpacedGrid(ContinuousGrid):
     @property
     def n_points(self) -> ScalarInt:
         """Return the total number of points in the grid."""
-        return self._piece_n_points.sum()
+        return self._piece_n_points.sum(dtype=jnp.int32)
 
     def to_jax(self) -> Float1D:
         """Convert the grid to a Jax array."""
@@ -157,7 +157,7 @@ class PiecewiseLogSpacedGrid(ContinuousGrid):
     @property
     def n_points(self) -> ScalarInt:
         """Return the total number of points in the grid."""
-        return self._piece_n_points.sum()
+        return self._piece_n_points.sum(dtype=jnp.int32)
 
     def to_jax(self) -> Float1D:
         """Convert the grid to a Jax array."""
@@ -243,8 +243,10 @@ def _init_piecewise_grid_cache(
     # Breakpoints are the effective starts of pieces 1..k-1
     breakpoints = starts[1:] if len(starts) > 1 else jnp.array([])
 
-    n_points = jnp.array([p.n_points for p in grid.pieces])
-    cumulative = jnp.concatenate([jnp.array([0]), jnp.cumsum(n_points[:-1])])
+    n_points = jnp.array([p.n_points for p in grid.pieces], dtype=jnp.int32)
+    cumulative = jnp.concatenate(
+        [jnp.array([0], dtype=jnp.int32), jnp.cumsum(n_points[:-1])]
+    )
 
     object.__setattr__(grid, "_breakpoints", breakpoints)
     object.__setattr__(grid, "_piece_starts", starts)
